Renamed JsonValueConverterNonGenericEnum to just JsonValueConverterEnum
