+++ src/System.Text.Json/src/System.Text.Json.csproj
@@ -55,7 +55,6 @@
     <Compile Include="System\Text\Json\Serialization\Converters\JsonValueConverterInt16.cs" />
     <Compile Include="System\Text\Json\Serialization\Converters\JsonValueConverterInt32.cs" />
     <Compile Include="System\Text\Json\Serialization\Converters\JsonValueConverterInt64.cs" />
-    <Compile Include="System\Text\Json\Serialization\Converters\JsonValueConverterNonGenericEnum.cs" />
     <Compile Include="System\Text\Json\Serialization\Converters\JsonValueConverterSByte.cs" />
     <Compile Include="System\Text\Json\Serialization\Converters\JsonValueConverterSingle.cs" />
     <Compile Include="System\Text\Json\Serialization\Converters\JsonValueConverterString.cs" />
